libに定義していたメソッドをモデルへ移動
